refactor(deploy): Port the forge client from Gitea to Forgejo (WIP)
Work in progress — the tree does not type-check yet. Committed with
--no-verify to preserve a large mechanical step; the branch is not
proposed for review in this state.

- gitea.py (2022 lines) becomes the Forgejo client inside forgejo.py.
  The three EXIT_* constants and render_ready_preamble from #676 are
  preserved because forgejo_runner.py imports them.
- The #676 account-only phase is removed: the ported
  run_configure_forgejo is a strict superset of it, so one
  forgejo-configure phase replaces two.
- Workspace repo is now nexus-<slug>-workspace, named for what it is
  rather than for the forge that happens to host it.
- Port 3200 -> 3202 throughout (3201 is git-proxy).
- Renamed the twelve modules whose gitea references were purely the
  internal role.

Remaining: __main__, and the five modules where both forges appear
(setup, service_env, s3_restore, infisical, config) — Gitea stays as a
normal opt-in stack, so those need selective edits. Then tests (4178
lines in test_gitea.py alone), compose, tofu, services.yaml, docs.
